Rework frame start and end calculations.
Correctly pull the timebase from the codec
configuration and use that for start and
end time computation.

Modify Rational class to handle more math.

+++ JNI/com/google/libvpx/IVFWriter.java
@@ -16,8 +16,8 @@ public class IVFWriter {
   private int fourcc;
   private int width;
   private int height;
-  private int rate;
-  private int scale;
+  private long rate;
+  private long scale;
   private int frameCount;
 
   public IVFWriter(File file,
